update database store operation from cast to call
Due to the poolboy implementation the database module lost the feature of all requests of the same list being handled by the same worker.

Because of it, the worker store operation must be updated from cast to call, otherwise the Todo.Server could issue a new write request on while the previous write request is not complete.

This could lead multiple workers to try access the same file at the same time, leading to race condition errors.